requirements.txt
exif==1.6.0
numpy==2.1.3
plum-py==0.8.7
pycountry==24.6.1
PyExifTool==0.5.6
reverse_geocode==1.6.5
scipy==1.14.1
python-dotenv~=1.0.1